Finished function (k), small reworks for reusability
Function k allows you to register a payment for a late fee. You can only do this on books that have been returned.
Also, every time the program asks for input, the user should be able to type exit to go back (except on the main menu where the user should select 12 to quit).
Also added a couple more safety nets for missing data in ResultSets.